+++ history_test.go
@@ -1,13 +1,13 @@
 package relic
 
 import (
-	"testing"
-
 	"fmt"
+	"testing"
+	"text/template"
 )
 
-func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
-	_, err := NewHistory("Test").RecordReleases(
+func TestHistory_DeclareReleases(t *testing.T) {
+	_, err := NewHistory("Test").DeclareReleases(
 		Release{
 			Version: parseVersion(t, "2.1.1"),
 			Notes:   `Everything fixed`,
@@ -23,7 +23,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 		t.Errorf("error expected")
 	}
 
-	history, err := NewHistory("Test").RecordReleases(
+	history, err := NewHistory("Test").DeclareReleases(
 		Release{
 			Version: parseVersion(t, "2.1.1"),
 			Notes:   `Everything fixed`,
@@ -51,7 +51,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 		changelog)
 
 	// Fail gap
-	_, err = NewHistory("Test").RecordReleases(
+	_, err = NewHistory("Test").DeclareReleases(
 		Release{
 			Version: parseVersion(t, "1.0.3"),
 			Notes:   `Wonderful things were achieved`,
@@ -67,7 +67,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 		t.Errorf("error expected")
 	}
 
-	history, err = NewHistory("Test").RecordReleases(
+	history, err = NewHistory("Test").DeclareReleases(
 		Release{
 			Version: parseVersion(t, "1.0.3"),
 			Notes:   `Wonderful things were achieved`,
@@ -93,7 +93,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 	assertChangelog(t, "# Test Changelog\n## Version 1.0.3\nWonderful things were achieved\n\n## Version 1.0.2\nHotfix\n\n## Version 1.0.1\nHotfix\n\n## Version 1.0.0\nWonderful things were achieved\n\n## Version 0.0.1\nMarvelous advances were made\n",
 		changelog)
 
-	_, err = NewHistory("Test").RecordReleases(
+	_, err = NewHistory("Test").DeclareReleases(
 		"0.1.3",
 		`Wonderful things were achieved`,
 		"0.0.2",
@@ -105,7 +105,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 		t.Errorf("error expected")
 	}
 
-	history, err = NewHistory("Test").RecordReleases(
+	history, err = NewHistory("Test").DeclareReleases(
 		"0.0.3",
 		`Wonderful things were achieved`,
 		"0.0.2",
@@ -123,7 +123,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 	assertChangelog(t, "# Test Changelog\n## Version 0.0.3\nWonderful things were achieved\n\n## Version 0.0.2\nWonderful things were achieved\n\n## Version 0.0.1\nMarvelous advances were made\n",
 		changelog)
 
-	_, err = NewHistory("Test").RecordReleases(
+	_, err = NewHistory("Test").DeclareReleases(
 		"0.0.3",
 		`Wonderful things were achieved`,
 		"0.0.2",
@@ -134,7 +134,7 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 		t.Errorf("error expected")
 	}
 
-	_, err = NewHistory("Test").RecordReleases(
+	_, err = NewHistory("Test").DeclareReleases(
 		"0.0.2",
 		`Wonderful things were achieved`,
 		"0.0.3",
@@ -147,9 +147,8 @@ func TestRegisterReleasesMonotonicAndUnique(t *testing.T) {
 	}
 }
 
-func TestMultipleRecordReleases(t *testing.T) {
-
-	history, err := NewHistory("Test").RecordReleases(
+func TestHistory_DeclareReleases_Multiple(t *testing.T) {
+	history, err := NewHistory("Test").DeclareReleases(
 		"0.1.0",
 		"Basic functionality",
 		"0.0.2",
@@ -167,7 +166,7 @@ func TestMultipleRecordReleases(t *testing.T) {
 	assertChangelog(t, "# Test Changelog\n## Version 0.1.0\nBasic functionality\n\n## Version 0.0.2\nBuild scripts\n\n## Version 0.0.1\nProof of concept\n",
 		changelog)
 
-	history1, err := history.RecordReleases(
+	history1, err := history.DeclareReleases(
 		"1.0.0",
 		"finally",
 		"0.2.1",
@@ -189,7 +188,7 @@ func TestMultipleRecordReleases(t *testing.T) {
 	assertChangelog(t, "# Test Changelog\n## Version 1.0.0\nfinally\n\n## Version 0.2.1\nPatch\n\n## Version 0.2.0\nCame after 0.1.0\n\n## Version 0.1.0\nBasic functionality\n\n## Version 0.0.2\nBuild scripts\n\n## Version 0.0.1\nProof of concept\n",
 		changelog)
 
-	_, err = history.RecordReleases(
+	_, err = history.DeclareReleases(
 		"0.1.3",
 		`New newness`,
 		"0.1.2",
@@ -199,6 +198,52 @@ func TestMultipleRecordReleases(t *testing.T) {
 	)
 	if err == nil {
 		t.Errorf("error expected")
+	}
+}
+
+func TestHistory_WithChangelogTemplate(t *testing.T) {
+	history, err := NewHistory("Test Project").
+		WithChangelogTemplate(template.Must(template.New("tests").
+			Parse("{{range .Releases}}{{$.Name}} (v{{.Version}}): {{.Notes}}\n{{end}}"))).
+		DeclareReleases(
+			"0.1.0",
+			"Basic functionality",
+			"0.0.2",
+			"Build scripts",
+			"0.0.1",
+			"Proof of concept",
+		)
+	if err != nil {
+		t.Fatal(err)
+	}
+	changelog, err := history.Changelog()
+	if err != nil {
+		t.Error(err)
+	}
+	assertChangelog(t, "Test Project (v0.1.0): Basic functionality\nTest Project (v0.0.2): Build scripts\nTest Project (v0.0.1): Proof of concept\n",
+		changelog)
+}
+
+func TestHistory_Release(t *testing.T) {
+	history, err := NewHistory("Test Project").
+		DeclareReleases(
+			"0.1.0",
+			"Basic functionality",
+			"0.0.2",
+			"Build scripts",
+			"0.0.1",
+			"Proof of concept",
+		)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	release, err := history.Release("0.0.2")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if release.Notes != "Build scripts" {
+		t.Errorf("release notes should be 'Build scripts' but is '%s'", release.Notes)
 	}
 }