+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Relic Changelog
+## Version 1.0.0
+Minor improvements:
+- Rename DeclareReleases to DeclareReleases (breaking API change)
+- Add sample snippet to readme
+- Sign version tags
+
+
 ## Version 0.0.1
 First release of Relic extracted from various initial projects, it can:
 - Generate changelogs